+++ Pokemon/GameFadeInOut.cpp
@@ -10,7 +10,7 @@ void GameFadeInOut::Draw(Graphics & g)
 {
 	gm->gamePlay->Draw(g);
 
-	int tp = 255 * ((FADETIME - fadeTimer) / (FADETIME));
+	int tp = 255 * int((FADETIME - fadeTimer) / (FADETIME));
 	SolidBrush brush(Color(tp, 0, 0, 0));
 	g.FillRectangle(&brush, 0, 0, SCREEN_SIZE_X, SCREEN_SIZE_Y);
 }
